Added solution to CarTest of exercise3.
Please provide the main method with your factory in line 55. Otherwise
the class will not run.
A solution to the first part of the exercise will not be released, as
this was more of a creative exercise and there is not "one solution". If
you have trouble with the first part, please feel free to contact me and
we can have a look together.